Reorganized lib folder structure
