+++ lib/get-urls.js
@@ -1,8 +1,8 @@
 'use strict'
 
 module.exports = async (options) => {
-  const urlsFromCrawler = require('./get-urls-from-crawler')
-  const urlsFromFile = require('./get-urls-from-file')
+  const urlsFromCrawler = require('./helpers-get-urls/get-urls-from-crawler')
+  const urlsFromFile = require('./helpers-get-urls/get-urls-from-file')
 
   var data
   if (options.file === false) {
