+++ lib/helpers-validate-pages/get-result.js
