+++ lib/validate-pages.js
@@ -14,13 +14,13 @@ module.exports = async (pagesToValidate, options) => {
   const path = require('path')
   const pagesHash = require('crypto').createHash('sha1').update(pagesToValidate.join('_')).digest('hex')
   const clc = require('cli-color')
-  const Cache = require('../Cache')
-  const getResult = require('./get-result')
-  const printError = require('./print-error')
-  const printFinish = require('./print-finish')
-  const isNotFound = require('./is-not-found')
-  const isValidHtml = require('./is-valid-html')
-  const failFast = require('./fail-fast')
+  const Cache = require('./Cache')
+  const getResult = require('./helpers-validate-pages/get-result')
+  const printError = require('./helpers-validate-pages/print-error')
+  const printFinish = require('./helpers-validate-pages/print-finish')
+  const isNotFound = require('./helpers-validate-pages/is-not-found')
+  const isValidHtml = require('./helpers-validate-pages/is-valid-html')
+  const failFast = require('./helpers-validate-pages/fail-fast')
 
   var greenOnBlack = clc.xterm(47).bgXterm(0)
   var redOnBlack = clc.xterm(196).bgXterm(0)
